requirements.txt
flask
openai>=1.0.0
feedparser
newspaper3k
lxml_html_clean
python-dotenv
